create a standard to project
