+++ diary-app/diary_app/controllers/create_fact_controllers.py
@@ -1,0 +1,14 @@
+from fastapi import APIRouter
+
+from models.create_fact_models import Create_Fact
+
+router = APIRouter()
+
+database = []
+
+
+@router.post("/")
+async def create_fact(new_fact: Create_Fact):
+    database.append(new_fact)
+    response_body = new_fact
+    return response_body
