+++ diary-app/diary_app/controllers/list_all_facts_controllers.py
@@ -1,0 +1,11 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+database = [{"message": "Not database online"}]
+
+
+@router.get("/")
+async def list_all_facts():
+    response_body = database
+    return response_body
